+++ test/use_focus_node_test.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 
@@ -36,6 +37,28 @@ void main() {
       // ignore: invalid_use_of_protected_member
       () => focusNode.hasListeners,
       throwsAssertionError,
+    );
+  });
+
+  testWidgets('debugFillProperties', (tester) async {
+    await tester.pumpWidget(
+      HookBuilder(builder: (context) {
+        useFocusNode();
+        return const SizedBox();
+      }),
+    );
+
+    final element = tester.element(find.byType(HookBuilder));
+
+    expect(
+      element
+          .toDiagnosticsNode(style: DiagnosticsTreeStyle.offstage)
+          .toStringDeep(),
+      equalsIgnoringHashCodes(
+        'HookBuilder\n'
+        ' │ useFocusNode: FocusNode#00000\n'
+        ' └SizedBox(renderObject: RenderConstrainedBox#00000)\n',
+      ),
     );
   });