+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [Unreleased]
+
+- Fixed a bug where a `useProvider` from riverpod potentially did not rebuild widgets
+
 ## 0.13.2
 
 - Fixed a bug where on hot-reload, a `HookWidget` could potentailly not rebuild